tailchasing/semantic/performance.py
"""
Performance optimizations for hypervector operations.

Key optimizations:
- Bit-packed storage for binary vectors
- SIMD operations for similarity computation
- Approximate nearest neighbor search for large codebases
- Parallel encoding for multiple functions
"""

import numpy as np
from typing import List, Tuple
import numba
from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
import faiss
import pickle
import lz4.frame


class OptimizedHVSpace:
    """Performance-optimized hypervector space implementation."""
    
    def __init__(self, dim: int = 8192, bipolar: bool = True, seed: int = 42):
        self.dim = dim
        self.bipolar = bipolar
        self.seed = seed
        
        # Use bit-packed storage for binary vectors
        self.bits_per_chunk = 64
        self.n_chunks = (dim + self.bits_per_chunk - 1) // self.bits_per_chunk
        
        # Initialize FAISS index for ANN search
        self.faiss_index = None
        self._init_faiss_index()
    
    def _init_faiss_index(self):
        """Initialize FAISS index for approximate nearest neighbor search."""
        if self.bipolar:
            # Use LSH for bipolar vectors
            self.faiss_index = faiss.IndexLSH(self.dim, self.dim // 16)
        else:
            # Use binary index for binary vectors
            self.faiss_index = faiss.IndexBinaryFlat(self.dim // 8)
    
    @numba.jit(nopython=True)
    def _hamming_distance_fast(self, a: np.ndarray, b: np.ndarray) -> float:
        """Optimized Hamming distance using numba."""
        return np.sum(a != b) / len(a)
    
    def encode_batch(self, vectors: List[np.ndarray], n_workers: int = 4) -> List[np.ndarray]:
        """Encode multiple vectors in parallel."""
        with ThreadPoolExecutor(max_workers=n_workers) as executor:
            return list(executor.map(self._encode_single, vectors))
    
    def similarity_matrix(self, vectors: List[np.ndarray]) -> np.ndarray:
        """Compute full similarity matrix efficiently."""
        n = len(vectors)
        matrix = np.zeros((n, n))
        
        # Vectorized computation for bipolar
        if self.bipolar:
            stacked = np.stack(vectors)
            matrix = np.dot(stacked, stacked.T) / self.dim
        else:
            # Use bit operations for binary
            for i in range(n):
                for j in range(i, n):
                    sim = 1.0 - self._hamming_distance_fast(vectors[i], vectors[j])
                    matrix[i, j] = matrix[j, i] = sim
        
        return matrix
    
    def find_nearest_neighbors(self, query: np.ndarray, k: int = 10) -> List[Tuple[int, float]]:
        """Find k nearest neighbors using FAISS."""
        query_reshaped = query.reshape(1, -1).astype(np.float32)
        distances, indices = self.faiss_index.search(query_reshaped, k)
        
        return list(zip(indices[0], distances[0]))


class CompressedSemanticCache:
    """Compressed storage for semantic vectors and metadata."""
    
    def __init__(self, cache_dir: str):
        self.cache_dir = cache_dir
        self.compression_level = 3  # Balance speed vs size
    
    def save_vectors(self, vectors: Dict[str, np.ndarray], metadata: Dict[str, Any]):
        """Save vectors with LZ4 compression."""
        data = {
            'vectors': vectors,
            'metadata': metadata
        }
        
        serialized = pickle.dumps(data)
        compressed = lz4.frame.compress(serialized, compression_level=self.compression_level)
        
        cache_file = Path(self.cache_dir) / "semantic_vectors.lz4"
        cache_file.write_bytes(compressed)
    
    def load_vectors(self) -> Tuple[Dict[str, np.ndarray], Dict[str, Any]]:
        """Load compressed vectors."""
        cache_file = Path(self.cache_dir) / "semantic_vectors.lz4"
        if not cache_file.exists():
            return {}, {}
        
        compressed = cache_file.read_bytes()
        serialized = lz4.frame.decompress(compressed)
        data = pickle.loads(serialized)
        
        return data['vectors'], data['metadata']


class ParallelEncoder:
    """Parallel encoding for large codebases."""
    
    def __init__(self, space: OptimizedHVSpace, n_processes: int = None):
        self.space = space
        self.n_processes = n_processes or os.cpu_count()
    
    def encode_files(self, files: List[Path]) -> Dict[str, List[np.ndarray]]:
        """Encode all functions in files using process pool."""
        with ProcessPoolExecutor(max_workers=self.n_processes) as executor:
            # Split files into chunks
            chunk_size = max(1, len(files) // self.n_processes)
            chunks = [files[i:i + chunk_size] for i in range(0, len(files), chunk_size)]
            
            # Process chunks in parallel
            results = executor.map(self._encode_chunk, chunks)
            
            # Merge results
            all_vectors = {}
            for chunk_result in results:
                all_vectors.update(chunk_result)
            
            return all_vectors
    
    def _encode_chunk(self, files: List[Path]) -> Dict[str, List[np.ndarray]]:
        """Encode a chunk of files."""
        vectors = {}
        for file in files:
            # Parse and encode file
            # ... encoding logic ...
            pass
        return vectors
