pyproject.toml
[project]
name = "tail-chasing-detector"
version = "0.1.0"
description = "Detects LLM-assisted tail-chasing anti-patterns in Python codebases using semantic analysis."
authors = [{name="Rohan Vinaik", email="your.email@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["linting", "code-quality", "static-analysis", "llm", "semantic-analysis", "hypervectors"]

dependencies = [
    "typing-extensions>=4.5",
    "tomli; python_version<'3.11'",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "numpy>=1.21.0,<2.0",
    "scipy>=1.7.0,<2.0",
    "networkx>=3.0",
    "pathspec>=0.11.0",
    "chardet>=4.0",
]

[project.urls]
"Homepage" = "https://github.com/rohanvinaik/TailChasingFixer"
"Bug Reports" = "https://github.com/rohanvinaik/TailChasingFixer/issues"
"Source" = "https://github.com/rohanvinaik/TailChasingFixer"
"Documentation" = "https://tail-chasing-detector.readthedocs.io"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "coverage>=7.0", 
    "ruff>=0.1.0",
    "mypy>=1.0",
    "black>=23.0",
    "isort>=5.12",
    "hypothesis>=6.100",
    "typer>=0.12",
    "requests>=2.31",
    "psutil>=5.9",
    "gitpython>=3.1",
    "pre-commit>=3.0"
]
visualization = [
    "plotly>=5.0",
    "matplotlib>=3.5",
    "seaborn>=0.12",
    "networkx>=3.0"
]
ml = [
    "scikit-learn>=1.3",
    "umap-learn>=0.5",
    "faiss-cpu>=1.7"
]
performance = [
    "numba>=0.57",
    "ray>=2.0",
    "dask>=2023.1"
]
cli = [
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "watchdog>=3.0"
]
lsp = [
    "pygls>=1.0",
    "lsprotocol>=2023.0"
]
enhanced = [
    "networkx>=3.0",
    "numpy>=1.21.0,<2.0",
    "scipy>=1.7.0,<2.0",
    "scikit-learn>=1.3"
]
all = [
    "tail-chasing-detector[visualization,ml,performance,cli,lsp,enhanced]"
]

[project.scripts]
tailchasing = "tailchasing.cli:main"
tailchasing-enhanced = "tailchasing.cli_enhanced:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
